# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: HysteriaX Management API
  version: 1.0.0
  description: Versioned API for the HysteriaX management service.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://{domain}/api/v1
    variables:
      domain:
        default: localhost
security:
  - BearerAuth: []
paths:
  "/overview":
    get:
      operationId: getOverview
      summary: Read full dashboard summary and actionable issues
      responses:
        '200':
          description: Monitoring overview
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/OverviewResponse"
        '400':
          "$ref": "#/components/responses/Error"
  "/overview/history":
    get:
      operationId: getOverviewHistory
      summary: Read rolling monitoring history (24h, 7d, 30d; legacy today)
      responses:
        '200':
          description: Monitoring overview
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/OverviewHistory"
        '400':
          "$ref": "#/components/responses/Error"
        '404':
          "$ref": "#/components/responses/Error"
      x-swift-include-optional-query: true
      parameters:
      - name: range
        description: Rolling elapsed time (24h, 7d, 30d), with hourly or 24-hour buckets anchored at window start. Legacy today uses local midnight.
        in: query
        required: true
        schema:
          type: string
          enum:
          - 24h
          - today
          - 7d
          - 30d
      - name: timezone
        in: query
        required: true
        schema:
          type: string
      - name: node_id
        in: query
        required: false
        schema:
          type: string
      - name: source
        in: query
        required: false
        schema:
          type: string
          enum:
          - users
          - network
  /nodes:
    get:
      operationId: listNodes
      summary: List nodes
      responses:
        '200':
          description: Node list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/NodeSummary' }
    post:
      operationId: createNode
      summary: Create node
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NodeCreate' }
      responses:
        '201':
          description: Created node and one-time node token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatedEntity' }
        '400': { $ref: '#/components/responses/Error' }
  /nodes/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getNode
      summary: Get node
      responses:
        '200':
          description: Node detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NodeDetail' }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      operationId: updateNode
      summary: Update node
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NodePatch' }
      responses:
        '200':
          description: Updated node
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NodeUpdateResponse' }
        '409': { $ref: '#/components/responses/Error' }
    delete:
      operationId: deleteNode
      summary: Delete node and queue managed remote uninstall
      parameters:
        - name: expected_revision
          in: query
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '202':
          description: Queued remote uninstall job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobReceipt' }
        '204': { description: Deleted }
        '409': { $ref: '#/components/responses/Error' }
  /nodes/{id}/usage:
    parameters:
      - $ref: '#/components/parameters/Id'
    put:
      operationId: updateNodeUsage
      summary: Correct or reset whole-host package usage without deploying the node
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NodeUsageUpdate' }
      responses:
        '200':
          description: Updated usage and node revision
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NodeUsageUpdateResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /nodes/{id}/ssh-test:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      operationId: testNodeSSH
      summary: Test node ssh
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Revision' }
      responses:
        '202':
          description: Queued SSH test job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobReceipt' }
        '409': { $ref: '#/components/responses/Error' }
  /nodes/{id}/deploy:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      operationId: deployNode
      summary: Deploy node
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Revision' }
      responses:
        '202':
          description: Queued deployment job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobReceipt' }
        '409': { $ref: '#/components/responses/Error' }
  /nodes/{id}/sync:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      operationId: syncNode
      summary: Sync node
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Revision' }
      responses:
        '202':
          description: Queued synchronization job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobReceipt' }
        '409': { $ref: '#/components/responses/Error' }
  /nodes/{id}/rollback:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      operationId: rollbackNode
      summary: Rollback node
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Revision' }
      responses:
        '202':
          description: Queued rollback job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobReceipt' }
        '409': { $ref: '#/components/responses/Error' }
  /nodes/{id}/resources:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: listNodeResources
      summary: List node resources
      responses:
        '200':
          description: Resource metadata without resource contents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/NodeResource' }
    post:
      operationId: uploadNodeResource
      summary: Upload a node resource
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResourceCreate' }
      responses:
        '201':
          description: Created encrypted resource and reference URI
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResourceReceipt' }
        '400': { $ref: '#/components/responses/Error' }
  /nodes/{id}/resources/{resource_id}:
    parameters:
      - $ref: '#/components/parameters/Id'
      - name: resource_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      operationId: deleteNodeResource
      summary: Delete an unreferenced node resource
      responses:
        '204': { description: Deleted resource }
        '409': { $ref: '#/components/responses/Error' }
  /users:
    get:
      operationId: listUsers
      summary: List users
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/UserSummary' }
    post:
      operationId: createUser
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreate' }
      responses:
        '201':
          description: Created user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatedEntity' }
  /users/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getUser
      summary: Get user
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSummary' }
    patch:
      operationId: updateUser
      summary: Update user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserPatch' }
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserUpdateResponse' }
        '409': { $ref: '#/components/responses/Error' }
    delete:
      operationId: deleteUser
      summary: Delete user
      parameters:
        - name: expected_revision
          in: query
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '204': { description: Deleted }
  /users/{id}/assignments:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      operationId: assignUserToNode
      summary: Assign user to node
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignmentRequest' }
      responses:
        '201':
          description: Assignment and one-time Hysteria credential
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssignmentReceipt' }
        '409': { $ref: '#/components/responses/Error' }
...
  /users/{id}/credentials/rotate:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      operationId: rotateUserCredentials
      summary: Rotate user credentials
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Revision' }
      responses:
        '200':
          description: "Rotated credentials, shown once"
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RotatedCredentials' }
...
  /users/{id}/usage:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getUserUsage
      summary: Get user usage
      responses:
        '200':
          description: Aggregated usage and freshness
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserUsageResponse' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Structured API error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/APIErrorResponse' }
  schemas:
...
    NodeUsageUpdateResponse:
      type: object
      required: [id, revision]
      properties:
        id: { type: string }
        revision: { type: integer }
...
    JobReceipt:
      type: object
      required: [job_id, status]
      properties:
        job_id: { type: string, format: uuid }
        status: { type: string }
...
    CreatedEntity:
      type: object
      properties:
        id: { type: string, format: uuid }
        revision: { type: integer }
        node_auth_token: { type: string, description: One-time node authentication token. }
        node: { $ref: '#/components/schemas/NodeReceipt' }
      additionalProperties: true
...
    OverviewHistory:
      type: object
      required:
      - generated_at
      - range
      - timezone
      - source
      - buckets
      properties:
        generated_at:
          type: string
        range:
          type: string
        timezone:
          type: string
        source:
          type: string
        buckets:
          type: array
          items:
            "$ref": "#/components/schemas/OverviewBucket"